requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
Babel==2.6.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bs4==0.0.1
captcha==0.3
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
diff-match-patch==20121119
Django==1.11
django-celery==3.2.2
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-reversion==2.0.13
django-simple-captcha==0.5.5
djangorestframework==3.8.2
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
flower==0.9.2
future==0.16.0
httpie==0.9.9
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==3.0.37
lxml==4.2.3
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.15.4
odfpy==1.3.6
openpyxl==2.5.4
pandas==0.23.4
Pillow==5.0.0
protobuf==3.6.0
pycparser==2.18
pycryptodomex==3.7.0
Pygments==2.2.0
PyMySQL==0.8.1
python-alipay-sdk==1.8.0
python-dateutil==2.7.5
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
tablib==0.12.1
tornado==5.1.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
